separate the sync process from opening a websocket connection
introducing an autoInitSync config property to initiate the sync process on a succesfull server connection, only when a previous timeoffset has not been established. This is usefull if the websocket connection is somehow broken mid flight, and we need to re-establish it, but don’t need to go thru the whole sync process again, because the timeoffset is likely not to have changed.